preliminary workbench importing.

+++ specifyweb/workbench/views.py
@@ -4,6 +4,7 @@ import errno
 import json
 import logging
 import subprocess
+import csv
 from glob import glob
 from uuid import uuid4
 
@@ -13,7 +14,7 @@ from django.views.decorators.csrf import csrf_exempt
 from django.db import connection, transaction
 from django.conf import settings
 
-from specifyweb.specify.api import toJson, get_object_or_404
+from specifyweb.specify.api import toJson, get_object_or_404, create_obj, obj_to_data
 from specifyweb.specify.views import login_maybe_required, apply_access_control
 from specifyweb.specify import models
 
@@ -60,16 +61,17 @@ def load(wb_id):
     rows = cursor.fetchall()
     return http.HttpResponse(toJson(rows), content_type='application/json')
 
-def save(wb_id, data):
+def save(wb_id, data, new=False):
     wb_id = int(wb_id)
     cursor = connection.cursor()
 
-    logger.debug("truncating wb %d", wb_id)
-    cursor.execute("""
-    delete wbdi from workbenchdataitem wbdi, workbenchrow wbr
-    where wbr.workbenchrowid = wbdi.workbenchrowid
-    and wbr.workbenchid = %s
-    """, [wb_id])
+    if not new:
+        logger.debug("truncating wb %d", wb_id)
+        cursor.execute("""
+        delete wbdi from workbenchdataitem wbdi, workbenchrow wbr
+        where wbr.workbenchrowid = wbdi.workbenchrowid
+        and wbr.workbenchid = %s
+        """, [wb_id])
 
     logger.debug("getting wb mapping items")
     cursor.execute("""
@@ -82,13 +84,17 @@ def save(wb_id, data):
     """, [wb_id])
 
     wbtmis = [r[0] for r in cursor.fetchall()]
-    assert len(wbtmis) + 1 == len(data[0]), (wbtmis, data[0])
+    assert len(wbtmis) + (0 if new else 1) == len(data[0]), (wbtmis, data[0])
 
-    logger.debug("clearing row numbers")
-    cursor.execute("update workbenchrow set rownumber = null where workbenchid = %s",
-                   [wb_id])
+    if new:
+        new_rows = [(i, wb_id) for i in range(len(data))]
+    else:
+        logger.debug("clearing row numbers")
+        cursor.execute("update workbenchrow set rownumber = null where workbenchid = %s",
+                       [wb_id])
 
-    new_rows = [(i, wb_id) for i, row in enumerate(data) if row[0] is None]
+        new_rows = [(i, wb_id) for i, row in enumerate(data) if row[0] is None]
+
     logger.debug("inserting %d new rows", len(new_rows))
     cursor.executemany("insert workbenchrow(rownumber, workbenchid) values (%s, %s)",
                        new_rows)
@@ -109,13 +115,14 @@ def save(wb_id, data):
         if row[0] is not None
     ])
 
-    logger.debug("removing deleted rows")
-    cursor.execute("""
-    delete from workbenchrow
-    where workbenchid = %s
-    and rownumber is null
-    """, [wb_id])
-    logger.debug("deleted %d rows", cursor.rowcount)
+    if not new:
+        logger.debug("removing deleted rows")
+        cursor.execute("""
+        delete from workbenchrow
+        where workbenchid = %s
+        and rownumber is null
+        """, [wb_id])
+        logger.debug("deleted %d rows", cursor.rowcount)
 
     logger.debug("inserting new wb values")
     cursor.executemany("""
@@ -123,9 +130,9 @@ def save(wb_id, data):
     (celldata, workbenchtemplatemappingitemid, workbenchrowid)
     values (%s, %s, %s)
     """, [
-        (celldata, wbtmi, new_row_id[i] if row[0] is None else row[0])
+        (celldata, wbtmi, new_row_id[i] if new or (row[0] is None) else row[0])
         for i, row in enumerate(data)
-        for wbtmi, celldata in zip(wbtmis, row[1:])
+        for wbtmi, celldata in zip(wbtmis, row if new else row[1:])
         if celldata is not None
     ])
     return load(wb_id)
@@ -218,3 +225,32 @@ def upload_status_list(request, wb_id):
     log_fnames = glob(os.path.join(settings.WB_UPLOAD_LOG_DIR, '%s_%s_*' % (settings.DATABASE_NAME, wb_id,)))
     statuses = [status_from_log(log) for log in log_fnames]
     return http.HttpResponse(toJson(statuses), content_type='application/json')
+
+@csrf_exempt
+@login_maybe_required
+@apply_access_control
+@require_POST
+@transaction.commit_on_success
+def import_workbench(request):
+    upload_file = request.FILES['file']
+
+    template = create_obj(request.specify_collection,
+                          request.specify_user_agent,
+                          'workbenchtemplate',
+                          json.loads(request.POST['template']))
+
+    template.name = upload_file.name
+    template.srcfilepath = upload_file.name
+    template.save()
+
+    workbench = models.Workbench.objects.create(
+        workbenchtemplate=template,
+        specifyuser=template.specifyuser,
+        name=template.name,
+        srcfilepath=template.name,
+    )
+
+    data = list(csv.reader(upload_file))[1:]
+    save(workbench.id, data, new=True)
+    return http.HttpResponse(toJson(obj_to_data(workbench)), status=201, content_type='application/json')
+
